Handle JoinedStr nodes (f-strings in Python 3)
This diff adds special handling for JoinedStr nodes (which are produced by f-strings added in Python 3.6). This handling mirrors what is already being done for normal strings.

I added a test that mirrors the existing Str test, saw that it failed, implemented the fix, and saw that it now succeeds. I ran the test on Python 2.7 and 3.6 and saw that it passed on both. I temporarily took out the guard at the beginning of the test and ran in on Python 2 and saw that it failed with a syntax error.