Adjust With nodes to include the "with " text
Apparently in Python 2 the `col_offset` of `With` nodes starts after the "with " text. This behavior is confusing and inconsistent with Python 3, which has the offset start at the beginning of the "with".

For an example of the issue, run the following commands:

```bash
$ python3 -c 'import ast; print(ast.parse("with foo: pass").body[0].col_offset)'
0
$ python -c 'import ast; print(ast.parse("with foo: pass").body[0].col_offset)'
5
```

This diff adds a test that the `With` node includes the "with" text, and corresponding behavior to fix up the start location when running under Python 3.