Explore a clearer lazy-parsing behaviour API
This introduces a new class which has the lazy-parsing behaviour,
separate from ASTTokens, thus leaving ASTTokens semantics unchanged.

ASTTokens still learns how to handle nodes unmarked, however it is
left somewhat cautious in the use of them. The new class instead
prefers them wherever it can.

While it looks like there's a lot of change here, there isn't --
most of the diff is moving files to the top of the file.