Remove 'unmarked' from public API
This was previously discouraged for use in general anyway and now
that we have a viable alternative there's no need for it.

The bulk of the diff here is moving the unmarked logic verbatim
down into the ASTText class as that is now the only place where
it is actually used.